+++ packages/xbell/CHANGELOG.md
@@ -1,5 +1,23 @@
 # xbell
 
+## 0.3.3
+
+### Patch Changes
+
+- e1152d9: handle nth first last api
+
+## 0.3.2
+
+### Patch Changes
+
+- 393a817: Support chained mode query
+
+## 0.3.1
+
+### Patch Changes
+
+- 532441c: ctx.expect support toMatchSnapshot
+
 ## 0.3.0
 
 ### Minor Changes